chore(accounts/service/edit): rename to `EditService`
